requirements.txt
requests==2.20.1
flask==1.0.2
fire==0.1.3
addict==2.2.0
gunicorn==19.9.0
attrs==18.2.0
whitenoise==4.1.2